ARM64/configs: ginkgo: Enable LLVM's ShadowCallStack
ShadowCallStack (SCS) is an LLVM instrumentation mode that protects
against return address overwrites (like stack buffer overflows) by
saving a function's return address to a separately allocated ShadowCallStack
in the function prolog of nonleaf functions and loading the return
address from the ShadowCallStack in the function epilog.
The return address is also stored on the regular stack for compatibility
with unwinders, but is otherwise unused. This ensures that attacks that
modify the return address on the regular stack have no effect on program
control flow.

Lookup:
SCS - https://source.android.com/devices/tech/debug/shadow-call-stack

Change-Id: I3fc42dcad9eb754510f4f0eb10caec8a752c43b2
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>