ARM64/configs: ginkgo: Enable kernel Control Flow Integrity (kCFI)
CFI is a security mechanism that disallows changes to the original
control flow graph of a compiled binary, making it significantlty
harder to perform any such attacks.

CFI is documented in -
[1] AOSP : https://source.android.com/devices/tech/debug/kcfi
[2] LWN.net : https://lwn.net/Articles/810077/
[3] Red Hat : https://www.redhat.com/en/blog/fighting-exploits-control-flow-integrity-cfi-clang
[4] Clang 12 docs : https://clang.llvm.org/docs/ControlFlowIntegrity.html

GrapheneOS has made CFI_CLANG a requirement in Android R. Sourced in [5].

[5] https://github.com/GrapheneOS/kernel_configs/commit/a227f7b86a9527d2717bf72dee03bfbdd7e2375e

CFI_CLANG requires LTO_CLANG to be enabled.

Test: Normal CTS test passes ( test proposed in AOSP Docs )
Change-Id: Ia6e98853386eb3003b6f5af3cda5c49019ce53e9
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>